refactor: table과 tablegroup 패키지 통합
