+++ src/main/java/kitchenpos/table/domain/TableGroupRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.tablegroup.domain;
+package kitchenpos.table.domain;
 
-import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
